fixed comment section updates
