+++ frontend/src/components/CommentSection/CommentBtns.js
@@ -1,17 +1,19 @@
-
-
-export default function CommentBtns ({ sessionUser, comment, setShowEdit, setEditingComment, setUpdatedComment, handleDelete }) {
+export default function CommentBtns ({ sessionUser, comment, setShowEdit, handleDelete }) {
 
     let commentBtns = null;
     if (sessionUser && sessionUser.id === comment.user_id) {
         commentBtns = (
             <div>
-                <button type='button' className='comment-btns' onClick={() => {
-                    setShowEdit(true);
-                    setEditingComment(comment.id);
-                    setUpdatedComment(comment.comment)
-                }}>Edit</button>
-                <button type='button' className='comment-btns' onClick={() => handleDelete(comment.id)}>Delete</button>
+                <button
+                    type='button'
+                    className='comment-btns'
+                    onClick={() => setShowEdit(true)}
+                >Edit</button>
+                <button
+                    type='button'
+                    className='comment-btns'
+                    onClick={handleDelete}
+                >Delete</button>
             </div>
         )
     } else {
